+++ src/CfdiUtils/Cleaner/CleanerException.php
@@ -1,0 +1,6 @@
+<?php
+namespace CfdiUtils\Cleaner;
+
+class CleanerException extends \RuntimeException
+{
+}