+++ tests/CfdiUtilsTests/Elements/Cfdi33/ComprobanteTest.php
@@ -1,6 +1,7 @@
 <?php
 namespace CfdiUtilsTests\Elements\Cfdi33;
 
+use CfdiUtils\Elements\Cfdi33\Addenda;
 use CfdiUtils\Elements\Cfdi33\CfdiRelacionado;
 use CfdiUtils\Elements\Cfdi33\CfdiRelacionados;
 use CfdiUtils\Elements\Cfdi33\Complemento;
@@ -8,6 +9,7 @@ use CfdiUtils\Elements\Cfdi33\Comprobante;
 use CfdiUtils\Elements\Cfdi33\Concepto;
 use CfdiUtils\Elements\Cfdi33\Conceptos;
 use CfdiUtils\Elements\Cfdi33\Emisor;
+use CfdiUtils\Elements\Cfdi33\Impuestos;
 use CfdiUtils\Elements\Cfdi33\Receptor;
 use CfdiUtils\Nodes\Node;
 use PHPUnit\Framework\TestCase;
@@ -119,6 +121,25 @@ class ComprobanteTest extends TestCase
         $this->assertSame($addReturn, $this->element);
     }
 
+    public function testGetAddenda()
+    {
+        $this->assertNull($this->element->searchNode('cfdi:Addenda'));
+        $child = $this->element->getAddenda();
+        $this->assertInstanceOf(Addenda::class, $child);
+        $this->assertSame($child, $this->element->searchNode('cfdi:Addenda'));
+    }
+
+    public function testAddAddenda()
+    {
+        $this->assertCount(0, $this->element);
+
+        $child = new Node('first');
+        $addReturn = $this->element->addAddenda($child);
+        $this->assertCount(1, $this->element);
+        $this->assertSame($child, $this->element->searchNode('cfdi:Addenda', 'first'));
+        $this->assertSame($addReturn, $this->element);
+    }
+
     public function testHasFixedAttributes()
     {
         $namespace = 'http://www.sat.gob.mx/cfd/3';
@@ -126,5 +147,26 @@ class ComprobanteTest extends TestCase
         $this->assertSame($namespace, $this->element['xmlns:cfdi']);
         $this->assertStringStartsWith($namespace . ' http://', $this->element['xsi:schemaLocation']);
         $this->assertNotEmpty($this->element['xmlns:xsi']);
+    }
+
+    public function testChildrenOrder()
+    {
+        // add in inverse order
+        $this->element->getAddenda();
+        $this->element->getComplemento();
+        $this->element->getImpuestos();
+        $this->element->getConceptos();
+        $this->element->getReceptor();
+        $this->element->getEmisor();
+        $this->element->getCfdiRelacionados();
+
+        // retrieve in correct order
+        $this->assertInstanceOf(CfdiRelacionados::class, $this->element->children()->get(0));
+        $this->assertInstanceOf(Emisor::class, $this->element->children()->get(1));
+        $this->assertInstanceOf(Receptor::class, $this->element->children()->get(2));
+        $this->assertInstanceOf(Conceptos::class, $this->element->children()->get(3));
+        $this->assertInstanceOf(Impuestos::class, $this->element->children()->get(4));
+        $this->assertInstanceOf(Complemento::class, $this->element->children()->get(5));
+        $this->assertInstanceOf(Addenda::class, $this->element->children()->get(6));
     }
 }