+++ tests/CfdiUtilsTests/Elements/Pagos10/DoctoRelacionadoTest.php
@@ -1,0 +1,22 @@
+<?php
+namespace CfdiUtilsTests\Elements\Pagos10;
+
+use CfdiUtils\Elements\Pagos10\DoctoRelacionado;
+use PHPUnit\Framework\TestCase;
+
+class DoctoRelacionadoTest extends TestCase
+{
+    /** @var DoctoRelacionado */
+    public $element;
+
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->element = new DoctoRelacionado();
+    }
+
+    public function testConstructedObject()
+    {
+        $this->assertSame('pagos10:DoctoRelacionado', $this->element->getElementName());
+    }
+}