+++ tests/CfdiUtilsTests/Nodes/NodesTest.php
@@ -112,6 +112,28 @@ class NodesTest extends TestCase
         $nodes->importFromArray([$specimen]);
     }
 
+    public function testGetThrowsExceptionWhenNotFound()
+    {
+        $nodes = new Nodes();
+        $this->expectException(\OutOfRangeException::class);
+        $this->expectExceptionMessage('The index 0 does not exists');
+        $nodes->get(0);
+    }
+
+    public function testGetWithExistentElements()
+    {
+        $foo = new Node('foo');
+        $bar = new Node('bar');
+        $nodes = new Nodes([$foo, $bar]);
+
+        $this->assertSame($foo, $nodes->get(0));
+        $this->assertSame($bar, $nodes->get(1));
+
+        // get after remove
+        $nodes->remove($foo);
+        $this->assertSame($bar, $nodes->get(0));
+    }
+
     public function testGetNodesByName()
     {
         $nodes = new Nodes();
@@ -131,5 +153,40 @@ class NodesTest extends TestCase
         $this->assertTrue($byName->exists($first));
         $this->assertTrue($byName->exists($second));
         $this->assertTrue($byName->exists($third));
+    }
+
+    public function testOrderedChildren()
+    {
+        $nodes = new Nodes([
+            new Node('foo'),
+            new Node('bar'),
+            new Node('baz'),
+        ]);
+        // test initial order
+        $this->assertEquals(
+            ['foo', 'bar', 'baz'],
+            [$nodes->get(0)->name(), $nodes->get(1)->name(), $nodes->get(2)->name()]
+        );
+
+        // sort previous values
+        $nodes->setOrder(['baz', '', '0', 'foo', '', 'bar', 'baz']);
+        $this->assertEquals(['baz', 'foo', 'bar'], $nodes->getOrder());
+        $this->assertEquals(
+            ['baz', 'foo', 'bar'],
+            [$nodes->get(0)->name(), $nodes->get(1)->name(), $nodes->get(2)->name()]
+        );
+
+        // add other baz (inserted at the bottom)
+        $nodes->add(new Node('baz', ['id' => 'second']));
+        $this->assertEquals(
+            ['baz', 'baz', 'foo'],
+            [$nodes->get(0)->name(), $nodes->get(1)->name(), $nodes->get(2)->name()]
+        );
+        $this->assertEquals('second', $nodes->get(1)['id']);
+
+        // add other not listed
+        $notListed = new Node('yyy');
+        $nodes->add($notListed);
+        $this->assertSame($notListed, $nodes->get(4));
     }
 }